user input punction in utils (needs to be defined before call)
